Users create endpoint working OK, at POST /api/users/0

+++ source/framework/ApiServiceHandler.py
@@ -29,7 +29,7 @@ class ApiServiceHandler(BaseHandler):
         if not user:
             response = BAD_AUTH_RESPONSE
         else:
-            response = method(user, url_id)
+            response = method(user, long(url_id))
 
         if 'status' in response.keys():
             self.response.set_status(response['status'])
